+++ my-new-app/src/redux/sagas/index.js
@@ -1,6 +1,15 @@
+import { all } from "redux-saga/effects";
 import { LoginSaga } from "./loginSaga";
 import { membersSaga } from "./membersSaga";
+import { projectsSaga } from "./projectsSaga";
+import { tasksSaga} from "./tasksSaga";
 
 export default function* rootSaga() {
-  yield LoginSaga();
+  yield all([
+    LoginSaga(),
+    membersSaga(),
+    projectsSaga(),
+    tasksSaga()
+  ])
+  
 }
